+++ tests/test-sqlite.R
@@ -7,17 +7,17 @@ connectionDetails <- DatabaseConnector::createConnectionDetails(
 
 workFolder <- tempfile("work")
 dir.create(workFolder)
-resultsFolder <- tempfile("results")
-dir.create(resultsFolder)
+testResultsFolder <- tempfile("results")
+dir.create(testResultsFolder)
 jobContext <- readRDS("tests/testJobContext.rds")
 jobContext$moduleExecutionSettings$workSubFolder <- workFolder
-jobContext$moduleExecutionSettings$resultsSubFolder  <- resultsFolder
+jobContext$moduleExecutionSettings$resultsSubFolder  <- testResultsFolder
 jobContext$moduleExecutionSettings$resultsConnectionDetails <- connectionDetails
 
 test_that("Run module", {
   source("Main.R")
   execute(jobContext)
-  resultsFiles <- list.files(resultsFolder)
+  resultsFiles <- list.files(testResultsFolder)
   expect_true("es_analysis.csv" %in% resultsFiles)
   expect_true("es_cm_result.csv" %in% resultsFiles)
   expect_true("es_cm_diagnostics_summary.csv" %in% resultsFiles)
@@ -27,19 +27,19 @@ test_that("Run module", {
 
 test_that("Skipped analyses as specified", {
   # We specified we didn't want cohort method analysis ID 2 in evidence synthesis ID 2:
-  results <- CohortGenerator::readCsv(file.path(resultsFolder, "es_cm_result.csv"))
+  results <- CohortGenerator::readCsv(file.path(testResultsFolder, "es_cm_result.csv"))
   expect_false(any(results$evidenceSynthesisAnalysisId == 2 & results$analysisId == 2))
 
-  results <- CohortGenerator::readCsv(file.path(resultsFolder, "es_sccs_result.csv"))
+  results <- CohortGenerator::readCsv(file.path(testResultsFolder, "es_sccs_result.csv"))
   expect_false(any(results$evidenceSynthesisAnalysisId == 2 & results$analysisId == 2))
 
 })
 
 test_that("Output conforms to results model", {
-  model <-  CohortGenerator::readCsv(file.path(resultsFolder, "resultsDataModelSpecification.csv"))
+  model <-  CohortGenerator::readCsv(file.path(testResultsFolder, "resultsDataModelSpecification.csv"))
   tables <- unique(model$tableName)
   for (table in tables) {
-    data <- readr::read_csv(file.path(resultsFolder, sprintf("%s.csv", table)), show_col_types = FALSE)
+    data <- readr::read_csv(file.path(testResultsFolder, sprintf("%s.csv", table)), show_col_types = FALSE)
     observed <- colnames(data)
     observed <- sort(observed)
     expected <- model$columnName[model$tableName == table]
@@ -73,13 +73,13 @@ test_that("Don't error when no negative controls present", {
   tempJobContext$moduleExecutionSettings$resultsConnectionDetails <- tempConnectionDetails
   execute(tempJobContext)
 
-  estimates <- readr::read_csv(file.path(resultsFolder, "es_cm_result.csv"), show_col_types = FALSE)
+  estimates <- readr::read_csv(file.path(testResultsFolder, "es_cm_result.csv"), show_col_types = FALSE)
   expect_gt(nrow(estimates), 0)
   expect_true(all(is.na(estimates$calibrated_rr)))
 })
 
 
-# readr::write_csv(OhdsiRTools::createResultsSchemaStub(resultsFolder), "resultsDataModelSpecification.csv")
+# readr::write_csv(OhdsiRTools::createResultsSchemaStub(testResultsFolder), "resultsDataModelSpecification.csv")
 
 unlink(workFolder)
-unlink(resultsFolder)
+unlink(testResultsFolder)